requirements.txt
cffi==1.8.3
pycparser==2.14
pyspotify==2.0.5
SimpleWebSocketServer==0.1.0
six==1.10.0
websocket-client==0.37.0
wheel==0.24.0
